rhcos-toolbox: use same options for non default images
Using a Fedora image instead of the RHEL support tools image for
toolbox was causing an error only when running from an `oc debug
node` pod.  There are no issues when running directly on RHEL
CoreOS.  Adding the podman `--pid=host` option fixes the error
but only when usding `podman run`.  There seems to be a difference
between `podman run` and `podman create`, `podman start`, and
`podman exec`.

There was also a difference on the second invocation of toolbox if
a container was already created.  The first invocation would use
the images' CMD and the second would use /bin/sh if no command was
passed in.  The second invocation will now use the images' CMD
when executing into the container for a more consistent experience.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1850230