rhcos-toolbox: check for an empty RUN label
`podman` returns `<no value>` when inspecting a label that does not
exist. That string is interpreted as the presence of a RUN label, so
we try to start the container with a non-existent RUN label. Use some
go formatting to return an empty string when the label is empty.

We've somehow avoided this until it was reported in RHBZ#2048789.

Co-authored-by: Jonathan Lebon <jonathan@jlebon.com>