remove existing containers when doing image pull
In the case where there is a new container image available or if
someone switches the container image they are using, there may be an
existing container created on disk. In this case, we should remove the
existing container, so the user enters a freshly created container
from the new container image.